+++ iOS_Wegg/iOS_Wegg/Screen/Login/View/PasswordRecoveryEmailView.swift
@@ -1,14 +1,14 @@
 //
-//  EmailLoginView.swift
+//  PasswordRecoveryEmailView.swift
 //  iOS_Wegg
 //
-//  Created by 이건수 on 2025.01.18.
+//  Created by 이건수 on 2025.01.24.
 //
 
 import UIKit
 
-class EmailLoginView: UIView {
-    
+class PasswordRecoveryEmailView: UIView {
+
     // MARK: - Init
     
     override init(frame: CGRect) {
@@ -28,7 +28,7 @@ class EmailLoginView: UIView {
     }
     
     private let mainLabel = UILabel().then {
-        $0.text = "이메일과 비밀번호를 입력해주세요"
+        $0.text = "인증 번호를 받을 이메일을 입력해주세요"
         $0.font = UIFont.LoginFont.title
         $0.textColor = .black
         $0.textAlignment = .left
@@ -41,41 +41,15 @@ class EmailLoginView: UIView {
         $0.textColor = UIColor.LoginColor.labelColor
     }
     
-    private let emailTextField = LoginTextField(
+    let emailTextField = LoginTextField(
         placeholder: "  wegg@email.com",
         type: .email
     )
     
-    private let passwordLabel = UILabel().then {
-        $0.text = "비밀번호"
-        $0.font = UIFont.LoginFont.label
-        $0.textColor = UIColor.LoginColor.labelColor
-    }
-        
-    private let passwordTextField = LoginTextField(
-        placeholder: "  비밀번호",
-        type: .password
-    )
-    
-    let findPasswordButton = UIButton().then {
-        $0.setTitle("비밀번호를 잊으셨나요?", for: .normal)
-        $0.titleLabel?.font = UIFont.LoginFont.title
-    }
-    
     let loginButton = LoginButton(
-        title: "로그인",
+        title: "인증 번호 보내기",
         backgroundColor: UIColor.customColor(.secondary)
     )
-    
-    var email: String? {
-        get { emailTextField.text }
-        set { emailTextField.text = newValue }
-    }
-    
-    var password: String? {
-        get { passwordTextField.text }
-        set { passwordTextField.text = newValue }
-    }
     
     // MARK: - Setup
     
@@ -85,9 +59,6 @@ class EmailLoginView: UIView {
             mainLabel,
             emailLabel,
             emailTextField,
-            passwordLabel,
-            passwordTextField,
-            findPasswordButton,
             loginButton
         ].forEach { addSubview($0) }
     }
@@ -102,7 +73,7 @@ class EmailLoginView: UIView {
         mainLabel.snp.makeConstraints { make in
             make.top.equalTo(backButton.snp.bottom).offset(22)
             make.leading.equalTo(backButton)
-            make.width.equalTo(188)
+            make.width.equalTo(208)
         }
         
         emailLabel.snp.makeConstraints { make in
@@ -116,25 +87,11 @@ class EmailLoginView: UIView {
             make.trailing.equalToSuperview().offset(-17)
         }
         
-        passwordLabel.snp.makeConstraints { make in
-            make.top.equalTo(emailTextField.snp.bottom).offset(22)
-            make.leading.equalTo(backButton)
-        }
-        
-        passwordTextField.snp.makeConstraints { make in
-            make.top.equalTo(passwordLabel.snp.bottom).offset(9)
-            make.leading.equalTo(backButton)
-            make.trailing.equalTo(emailTextField)
-        }
-        
-        findPasswordButton.snp.makeConstraints { make in
-            make.top.equalTo(passwordTextField.snp.bottom).offset(10)
-            make.trailing.equalTo(emailTextField)
-        }
-        
         loginButton.snp.makeConstraints { make in
             make.centerX.equalToSuperview()
             make.bottom.equalToSuperview().offset(-40)
         }
+        
     }
+
 }
